Add EmbeddingsAccuracy to metrics.py
EmbeddingsAccuracy is a custom metrics that behave like
MulticlassAccuracy but perform argmax(dim=-1) on targets.
So there is the implicit assumption that the correct labels correspond
to the entry in the target vector with the highest value (this is true
for similarities and probabilities embeddings).
This metrics must be used when `path_embeddings` is provided to a
dataloader. Otherwise an error will be thrown.