Add AlignmentAccuracy in metrics.py
AlignmentAccuracy is a metrics that must be used to calculate accuracy
for models that returns embeddings as outputs. If the model returns
logits or probabilities you can take the argmax to get the predictions
labels, but if the model return an embedding you must compare it to the
class embeddings and select the top_k similar embeddings (we use
cosine similarity as similarity score).